requirements.txt
nextcord==2.6.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
typer==0.12.3
python-dotenv==1.0.1
loguru==0.7.2
aiohttp==3.10.5
pydantic==2.9.2
pyjwt==2.9.0
pytest==8.3.3
httpx==0.28.0
black==24.10.0
flake8==7.1.1
isort==5.13.2
setuptools>=68.0.0